refactor(workflows): extract hooks + channel-approvals from activate (C3b)
Phase C3b of the workflows/index.ts split. Pulls the two remaining logic
blocks out of activate() into focused modules.

- lib/register-hooks.ts — registerWorkflowHooks(ctx) registers all 21
  ctx.hooks.register RPC/event hooks (the runtime/instance/gate surface +
  the notification-channel lookups).
- lib/channel-approvals.ts — wireChannelApprovals(ctx) rehydrates pending
  approval renders and subscribes to runtime-channel approval responses,
  driving approveGate/rejectGate for channel-decided gates. Returns the
  unsubscribe handle, which index.ts keeps for onShutdown.

activate() now reads as orchestration: setWorkflowPluginContext →
registerFileBackedContentType → loadDefaultWorkflows → notification wiring →
wireChannelApprovals → registerWorkflowHooks → health checks →
registerWorkflowExecTools. index.ts sheds ~28 now-unused imports and drops
from 472 to 288 lines.

Pure behavior-preserving relocation — hook registrations and the channel
handler are byte-identical.

Verified: typecheck + eslint + workflows suite (420/0) + full suite (5124/0)
+ binary build + isolated-home boot smoke (plugin Ready; hooks + channel
wiring active; /definitions + /instances + /gates/pending → 200).
